% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_md.R
\name{BC_md}
\alias{BC_md}
\title{Bias correction for MLE from multiple-dilution assay data}
\usage{
BC_md(tau, M, q, dilutions)
}
\arguments{
\item{tau}{Parameter}

\item{M}{Total number of wells originally sequenced with the QVOA.}

\item{q}{Proportion of p24-positive wells that underwent UDSA.}

\item{dilutions}{Vector of all dilution levels (in millions of cells per well).}
}
\value{
A vector of the same length as \code{lambda}
}
\description{
Bias correction for MLE from multiple-dilution assay data
}
