% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BC_sd.R
\name{BC_sd}
\alias{BC_sd}
\title{Bias correction for MLE from single-dilution assay data}
\usage{
BC_sd(lambda, M, q)
}
\arguments{
\item{lambda}{Vector of DVL-specific parameters.}

\item{M}{Total number of wells originally sequenced with the QVOA.}

\item{q}{Proportion of p24-positive wells that underwent UDSA.}
}
\value{
A vector of the same length as \code{lambda}
}
\description{
Bias correction for MLE from single-dilution assay data
}
