% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_fisher_md.R
\name{deriv_fisher_md}
\alias{deriv_fisher_md}
\title{Derivative of the Fisher information of the log-likelihood for IUPM from multiple-dilution assay data}
\usage{
deriv_fisher_md(tau, M, q, u)
}
\arguments{
\item{tau}{Parameter}

\item{M}{Total number of wells originally sequenced with the QVOA.}

\item{q}{Proportion of p24-positive wells that underwent UDSA.}

\item{u}{Vector of all dilution levels (in millions of cells per well).}
}
\value{
A matrix
}
\description{
Derivative of the Fisher information of the log-likelihood for IUPM from multiple-dilution assay data
}
