% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv_fisher_sd.R
\name{deriv_fisher_sd}
\alias{deriv_fisher_sd}
\title{Derivative of the Fisher information of the log-likelihood for IUPM from single-dilution assay data}
\usage{
deriv_fisher_sd(lambda, M, q)
}
\arguments{
\item{lambda}{Vector of DVL-specific parameters.}

\item{M}{Total number of wells originally sequenced with the QVOA.}

\item{q}{Proportion of p24-positive wells that underwent UDSA.}
}
\value{
A matrix
}
\description{
Derivative of the Fisher information of the log-likelihood for IUPM from single-dilution assay data
}
