% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_third_deriv_md.R
\name{exp_third_deriv_md}
\alias{exp_third_deriv_md}
\title{Expected value of the third derivative of the Fisher information of the log-likelihood for IUPM from multiple-dilution assay data}
\usage{
exp_third_deriv_md(tau, M, q, u)
}
\arguments{
\item{tau}{Parameter}

\item{M}{Total number of wells originally sequenced with the QVOA.}

\item{q}{Proportion of p24-positive wells that underwent UDSA.}

\item{u}{Vector of all dilution levels (in millions of cells per well).}
}
\value{
A matrix
}
\description{
Expected value of the third derivative of the Fisher information of the log-likelihood for IUPM from multiple-dilution assay data
}
