% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_third_deriv_sd.R
\name{exp_third_deriv_sd}
\alias{exp_third_deriv_sd}
\title{Expected value of the third derivative of the Fisher information of the log-likelihood for IUPM from single-dilution assay data}
\usage{
exp_third_deriv_sd(lambda, M, q)
}
\arguments{
\item{lambda}{Vector of DVL-specific parameters.}

\item{M}{Total number of wells originally sequenced with the QVOA.}

\item{q}{Proportion of p24-positive wells that underwent UDSA.}
}
\value{
A matrix
}
\description{
Expected value of the third derivative of the Fisher information of the log-likelihood for IUPM from single-dilution assay data
}
