% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_md.R
\name{fisher_md}
\alias{fisher_md}
\title{Fisher information of the log-likelihood for IUPM from multiple-dilution assay data}
\usage{
fisher_md(tau, M, q, dilutions)
}
\arguments{
\item{tau}{Parameter}

\item{M}{Total number of wells originally sequenced with the QVOA.}

\item{q}{Proportion of p24-positive wells that underwent UDSA.}

\item{dilutions}{Vector of all dilution levels (in millions of cells per well).}
}
\value{
A matrix
}
\description{
Fisher information of the log-likelihood for IUPM from multiple-dilution assay data
}
