% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SLDeepAssay_md.R
\name{fit_SLDeepAssay_md}
\alias{fit_SLDeepAssay_md}
\title{Maximum likelihood estimator (MLE) for multiple-dilution assay data}
\usage{
fit_SLDeepAssay_md(
  assay = NULL,
  dilutions = NULL,
  assay.summary,
  bc = NULL,
  maxit = 1e+06,
  lb = 1e-06,
  ub = Inf
)
}
\arguments{
\item{assay}{List of data frames with assay data from each dilution level. Default is \code{NULL}.}

\item{dilutions}{Vector of dilutions. Default is \code{NULL}.}

\item{assay.summary}{(Optional) Instead of supplying \code{assay}, supply a summary of assay results. This summary should contain one row per dilution level and the following columns: M (total number of wells), n (number of distinct viral lineages \[DVL\]), MN (number of p24-negative wells), m (number of deep sequenced wells), Y1,..., Yn (counts of wells positive for DVL i, (i = 1,...,n), and dilutions (dilution levels, in millions of cells per well).}

\item{bc}{Logical indicator for returning bias corrected MLE and CI. By default, the bias correction is not computed if n>40.}

\item{maxit}{The maximum number of iterations (passed to \code{optim}). Default is \code{maxit = 1E4}.}

\item{lb}{Lower-bound on the IUPM (passed to \code{optim}). Default is \code{lb = 1E-6}.}

\item{ub}{Upper-bound on the IUPM (passed to \code{optim}). Default is \code{ub = Inf}.}
}
\value{
Named list with the following slots:
\item{mle}{MLE}
\item{se}{Standard error for the MLE}
\item{ci}{95\% confidence interval for the MLE}
\item{mle.bc}{Bias-corrected MLE}
\item{ci.bc}{95\% confidence interval for the bias-corrected MLE}
}
\description{
Maximum likelihood estimator (MLE) for multiple-dilution assay data
}
