% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SLDeepAssay_sd_imperfect.R
\name{fit_SLDeepAssay_sd_imperfect}
\alias{fit_SLDeepAssay_sd_imperfect}
\title{Maximum likelihood estimator (MLE) for single-dilution imperfect assays data}
\usage{
fit_SLDeepAssay_sd_imperfect(
  assay_QVOA,
  assay_UDSA,
  u = 1,
  sens_QVOA = 1,
  spec_QVOA = 1,
  sens_UDSA = 1,
  spec_UDSA = 1,
  maxit = 10000,
  lb = 1e-06,
  ub = Inf
)
}
\arguments{
\item{assay_QVOA}{Assay data, with columns representing the wells.}

\item{assay_UDSA}{Assay data, with rows representing the distinct viral lineages (DVL) and columns representing the wells.}

\item{u}{Dilution level in millions of cells per well (a positive scalar). Default is \code{u = 1}.}

\item{sens_QVOA}{Sensitivity (i.e., true positive rate) for the QVOA (a scalar between 0 and 1). Default is \code{sens_QVOA = 1}.}

\item{spec_QVOA}{Specificity (i.e., true negative rate) for the QVOA (a scalar between 0 and 1). Default is \code{spec_QVOA = 1}.}

\item{sens_UDSA}{Sensitivity (i.e., true positive rate) for the UDSA (a scalar between 0 and 1). Default is \code{sens_UDSA = 1}.}

\item{spec_UDSA}{Specificity (i.e., true negative rate) for the UDSA (a scalar between 0 and 1). Default is \code{spec_UDSA = 1}.}

\item{maxit}{The maximum number of iterations (passed to \code{optim}). Default is \code{maxit = 1E4}.}

\item{lb}{Parameter lower bound (passed to \code{optim}). Default is \code{lb = 1E-6}.}

\item{ub}{Parameter upper bound (passed to \code{optim}). Default is \code{ub = Inf}.}

\item{corrected}{Logical, if \code{corrected = TRUE} the bias-corrected MLE will be returned. If \code{corrected = FALSE} the bias-corrected MLE will be not be returned. If \code{corrected = NULL}, the bias correction will be computed if here are <= 40 DVLs in \code{assay}. Default is \code{corrected = NULL}.}
}
\value{
A named list with the following slots:
\item{mle}{MLE}
\item{convergence}{An integer code for the convergence status (returned by \code{optim})}
\item{message}{A character string giving any additional information about the convergence status (returned by \code{optim})}
}
\description{
Maximum likelihood estimator (MLE) for single-dilution imperfect assays data
}
