% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EY.R
\name{get_EY}
\alias{get_EY}
\title{Expected value of Y}
\usage{
get_EY(M, q, lambda)
}
\arguments{
\item{M}{Total number of wells originally sequenced with the QVOA.}

\item{q}{Proportion of p24-positive wells that underwent UDSA.}

\item{lambda}{Vector of DVL-specific parameters.}
}
\value{
A scalar
}
\description{
Expected value of Y
}
