% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloglik_md.R
\name{gloglik_md}
\alias{gloglik_md}
\title{Gradient of the log-likelihood for IUPM from multiple-dilution assay data}
\usage{
gloglik_md(tau, assay.summary)
}
\arguments{
\item{tau}{Parameter}

\item{assay.summary}{Data}
}
\value{
A vector (of length \code{n})
}
\description{
Gradient of the log-likelihood for IUPM from multiple-dilution assay data
}
