% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloglik_sd_imperfect.R
\name{gloglik_sd_imperfect}
\alias{gloglik_sd_imperfect}
\title{Gradient of the log-likelihood for IUPM from single-dilution imperfect assay data}
\usage{
gloglik_sd_imperfect(l, cd)
}
\arguments{
\item{l}{Vector of DVL-specific parameters.}

\item{M}{Total number of wells originally sequenced with the QVOA.}

\item{MPstar}{Number of p24-positive wells.}

\item{mstar}{Number of p24-positive wells that underwent the UDSA.}

\item{y}{A vector of  DVL-specific infection counts.}
}
\value{
A vector
}
\description{
Gradient of the log-likelihood for IUPM from single-dilution imperfect assay data
}
