% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_md.R
\name{loglik_md}
\alias{loglik_md}
\title{Log-likelihood for IUPM from multiple-dilution assay data}
\usage{
loglik_md(tau, assay.summary)
}
\arguments{
\item{tau}{Parameter}

\item{assay.summary}{Data}
}
\value{
A scalar
}
\description{
Log-likelihood for IUPM from multiple-dilution assay data
}
