% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion_lrt.R
\name{negbin_gloglik_md}
\alias{negbin_gloglik_md}
\title{Gradient of the log-likelihood for IUPM from multiple-dilution assay data}
\usage{
negbin_gloglik_md(tau, gamma, assay_summary)
}
\arguments{
\item{tau}{Vector of DVL-specific IUPM parameters}

\item{gamma}{Dispersion parameter (a non-negative number)}

\item{assay_summary}{Summary of assay data from multiple dilutions}
}
\value{
A vector (of length \code{n})
}
\description{
Gradient of the log-likelihood for IUPM from multiple-dilution assay data
}
