% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_assay_md.R
\name{simulate_assay_md}
\alias{simulate_assay_md}
\title{Simulate multiple-dilution assay data}
\usage{
simulate_assay_md(M, tau, q, dilutions, remove_undetected = T, seed = NULL)
}
\arguments{
\item{M}{Total number of wells (a scalar).}

\item{tau}{Parameters}

\item{q}{Proportion of p24-positive wells that underwent UDSA (a scalar between 0 and 1).}

\item{dilutions}{Vector of dilution levels (in millions of cells per well).}

\item{remove_undetected}{Logical, if \code{remove_undetected = TRUE} (the default), then DVL which were not detected in any of the deep sequenced wells are deleted.}

\item{seed}{(Optional) An integer setting the random seed used to simulate the assay data. Default is \code{seed = NULL}.}
}
\value{
Named list with the following slots:
\item{any_DVL}{A \code{Mx1} vector containing indicators of overall (any DVL) infection across the wells.}
\item{DVL_specific}{Standard error for the MLE}
}
\description{
Simulate multiple-dilution assay data
}
