% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_assay_sd.R
\name{simulate_assay_sd}
\alias{simulate_assay_sd}
\title{Simulate single-dilution assay data}
\usage{
simulate_assay_sd(
  M,
  tau,
  q,
  u = 1,
  sens_QVOA = 1,
  spec_QVOA = 1,
  sens_UDSA = 1,
  spec_UDSA = 1,
  sequence_all = FALSE,
  k = Inf,
  remove_undetected = TRUE
)
}
\arguments{
\item{M}{Total number of wells (a scalar).}

\item{tau}{Mean counts of cells per million infected with each DVL (a vector). (Note: All elements in \code{tau} must be > 0.)}

\item{q}{Proportion of p24-positive wells that underwent UDSA (a scalar between 0 and 1).}

\item{u}{Dilution level in millions of cells per well (a positive scalar). Default is \code{dilution = 1}.}

\item{sens_QVOA}{Sensitivity (i.e., true positive rate) for the QVOA (a scalar between 0 and 1). Default is \code{sens_QVOA = 1}.}

\item{spec_QVOA}{Specificity (i.e., true negative rate) for the QVOA (a scalar between 0 and 1). Default is \code{spec_QVOA = 1}.}

\item{sens_UDSA}{Sensitivity (i.e., true positive rate) for the UDSA (a scalar between 0 and 1). Default is \code{sens_UDSA = 1}.}

\item{spec_UDSA}{Specificity (i.e., true negative rate) for the UDSA (a scalar between 0 and 1). Default is \code{spec_UDSA = 1}.}

\item{sequence_all}{(Optional) Logical, if \code{sequence_all = FALSE} (the default), then only p24-positive wells are considered for UDSA. If \code{sequence_all = TRUE} and \code{q = 1}, then all wells will undergo UDSA.}

\item{k}{(Optional) Overdispersion parameter (a positive number). Default is \code{k = Inf}, which corresponds to no overdispersion.}

\item{remove_undetected}{Logical, if \code{remove_undetected = TRUE} (the default), then DVL which were not detected in any of the deep sequenced wells are deleted.}
}
\value{
Named list with the following slots:
\item{any_DVL}{A vector containing overall (any DVL) infection indicators of across the wells.}
\item{DVL_specific}{A matrix of infection indicators with rows and columns representing the DVLs and wells, respectively.}
}
\description{
Simulate single-dilution assay data
}
