% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_SLDeepAssay_md.R
\name{simulate_SLDeepAssay_md}
\alias{simulate_SLDeepAssay_md}
\title{Simulate multiple-dilution assay data and fit new and existing methods to it.}
\usage{
simulate_SLDeepAssay_md(
  M,
  tau,
  q,
  u,
  remove_undetected = TRUE,
  optim_method = "BFGS"
)
}
\arguments{
\item{M}{Total number of wells at each dilution level (a vector of length D).}

\item{tau}{DVL-specific IUPMs (a vector of length \code{n}). (Note: All elements in \code{tau} must be > 0.)}

\item{q}{Proportions of p24-positive wells that underwent UDSA at each dilution level (a vector of length D).}

\item{u}{Vector of dilution levels in millions of cells per well (a vector of length D).}

\item{remove_undetected}{Logical, if \code{remove_undetected = TRUE} (the default), then DVL which were not detected in any of the deep sequenced wells across all dilution levels are deleted.
#' @param optim_method optimization method ("BFGS" or "L-BFGS-B")}
}
\value{
Named list with the following slots:
\item{assay_summary}{Simulated multiple-dilution assay data (deep sequenced) in summary form.}
\item{MLE_woUDSA}{Point estimate, standard error estimate, and 95\% confidence interval for the MLE (without deep sequencing information).}
\item{BCMLE_woUDSA}{Point estimate, standard error estimate, and 95\% confidence interval for the bias-corrected MLE (without deep sequencing information).}
\item{MLE_wUDSA}{Point estimate, standard error estimate, and 95\% confidence interval for the MLE (with deep sequencing information).}
\item{BCMLE_wUDSA}{Point estimate, standard error estimate, and 95\% confidence interval for the bias-corrected MLE (with deep sequencing information).}
\item{Message}{(Optional) Message describing whether the \code{Assay} needed to be re-simulated due to either none of the DVL being detected (\code{Message = 1}) or at least one DVL being detected in all deep-sequenced wells (\code{Message = 2}). Otherwise, \code{Message = 0}.}
}
\description{
Simulate multiple-dilution assay data and fit new and existing methods to it.
}
