% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SLDeepAssay_md.R
\name{fit_SLDeepAssay_md}
\alias{fit_SLDeepAssay_md}
\title{Maximum likelihood estimator (MLE) for multiple-dilution assay data}
\usage{
fit_SLDeepAssay_md(
  assay = NULL,
  u = NULL,
  assay_summary,
  corrected = NULL,
  maxit = 1e+06,
  lb = 1e-06,
  ub = Inf,
  optim_method = "BFGS"
)
}
\arguments{
\item{assay}{List of data frames with assay data from each dilution level. Default is \code{NULL}.}

\item{u}{Vector of dilution levels, in millions of cells per well. Default is \code{NULL}.}

\item{assay_summary}{(Optional) Instead of supplying \code{assay} and \code{u}, supply a summary of assay results in the form of a data frame. This summary should contain one row per dilution level and the following columns: M (total number of wells), n (number of distinct viral lineages \[DVL\]), MN (number of p24-negative wells), m (number of deep sequenced wells), Y1,..., Yn (counts of wells positive for DVL i, (i = 1,...,n), and u (dilution levels, in millions of cells per well).}

\item{corrected}{Logical, if \code{corrected = TRUE} the bias-corrected MLE will be returned. If \code{corrected = FALSE} the bias-corrected MLE will be not be returned. If \code{corrected = NULL}, the bias correction will be computed if here are <= 40 DVLs in \code{assay}. Default is \code{corrected = NULL}.}

\item{maxit}{The maximum number of iterations (passed to \code{optim}). Default is \code{maxit = 1E4}.}

\item{lb}{Lower-bound on the IUPM (passed to \code{optim}). Default is \code{lb = 1E-6}.}

\item{ub}{Upper-bound on the IUPM (passed to \code{optim}). Default is \code{ub = Inf}.}

\item{optim_method}{optimization method ("BFGS" or "L-BFGS-B")}
}
\value{
Named list with the following slots:
\item{mle}{MLE}
\item{se}{Standard error for the MLE}
\item{ci}{95\% confidence interval for the MLE}
\item{mle_bc}{Bias-corrected MLE}
\item{ci_bc}{95\% confidence interval for the bias-corrected MLE}
}
\description{
Maximum likelihood estimator (MLE) for multiple-dilution assay data
}
