% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_SLDeepAssay_sd.R
\name{fit_SLDeepAssay_sd}
\alias{fit_SLDeepAssay_sd}
\title{Maximum likelihood estimator (MLE) for single-dilution assay data}
\usage{
fit_SLDeepAssay_sd(
  assay,
  dilution = 1,
  M = NULL,
  n = NULL,
  MP = NULL,
  m = NULL,
  Y = NULL,
  corrected = FALSE,
  maxit = 10000,
  lb = 1e-06,
  ub = Inf
)
}
\arguments{
\item{assay}{Assay data, with rows representing the distinct viral lineages (DVL) and columns representing the wells. Any columns with \code{NA} values will be treated as positive but missing deep sequencing information.}

\item{dilution}{Scalar, representing the dilution for number of cells per well (in millions). Default is \code{dilution = 1}.}

\item{M}{(Optional) Instead of supplying \code{assay}, supply the total number of wells (a scalar). Default is \code{M = NULL}.}

\item{n}{(Optional) Instead of supplying \code{assay}, supply the total number of DVL detected (a scalar). Default is \code{n = NULL}.}

\item{MP}{(Optional) Instead of supplying \code{assay}, supply the total number of p24-positive wells (a scalar). Default is \code{MP = NULL}.}

\item{m}{(Optional) Instead of supplying \code{assay}, supply the total number of p24-positive wells that underwent deep sequencing (a scalar). Default is \code{m = NULL}.}

\item{Y}{(Optional) Instead of supplying \code{assay}, supply the numbers of wells (without missing data) that were infected with each DVL (a vector of length \code{n}). Default is \code{Y = NULL}.}

\item{corrected}{Logical, if \code{corrected = TRUE} the bias-corrected MLE will be returned. If there are <= 40 DVL in \code{assay}, default is \code{corrected = TRUE}; otherwise default is \code{corrected = FALSE}.}

\item{maxit}{The maximum number of iterations (passed to \code{optim}). Default is \code{maxit = 1E4}.}

\item{lb}{Lower-bound on the IUPM (passed to \code{optim}). Default is \code{lb = 1E-6}.}

\item{ub}{Upper-bound on the IUPM (passed to \code{optim}). Default is \code{ub = Inf}.}
}
\value{
A named list with the following slots:
\item{mle}{MLE}
\item{se}{Standard error for the MLE}
\item{ci}{95\% confidence interval for the MLE}
\item{mle.bc}{Bias-corrected MLE}
\item{ci.bc}{95\% confidence interval for the bias-corrected MLE}
}
\description{
Maximum likelihood estimator (MLE) for single-dilution assay data
}
