% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloglik_md.R
\name{gloglik_md}
\alias{gloglik_md}
\title{Gradient of the log-likelihood for IUPM from multiple-dilution assay data}
\usage{
gloglik_md(tau, assay_summary)
}
\arguments{
\item{tau}{Vector of DVL-specific IUPM parameters}

\item{assay_summary}{Summary of assay data from multiple dilutions}
}
\value{
A vector (of length \code{n})
}
\description{
Gradient of the log-likelihood for IUPM from multiple-dilution assay data
}
