% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_md.R
\name{loglik_md}
\alias{loglik_md}
\title{Log-likelihood for IUPM from multiple-dilution assay data}
\usage{
loglik_md(tau, assay_summary)
}
\arguments{
\item{tau}{Vector of DVL-specific IUPM parameters}

\item{assay_summary}{Summary of assay data from multiple dilutions}
}
\value{
A scalar value of the log-likelihood
}
\description{
Log-likelihood for IUPM from multiple-dilution assay data
}
