% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion_lrt.R
\name{negbin_gloglik_sd}
\alias{negbin_gloglik_sd}
\title{Gradient of the negative binomial log-likelihood for IUPM from single-dilution assay data (with respect to tau and gamma)}
\usage{
negbin_gloglik_sd(l, gamma, M, MP, m, Y)
}
\arguments{
\item{l}{Vector of DVL-specific parameters.}

\item{gamma}{Dispersion parameter (a non-negative number)}

\item{M}{Total number of wells originally sequenced with the QVOA.}

\item{MP}{Number of p24-positive wells.}

\item{m}{Number of p24-positive wells that underwent the UDSA.}

\item{y}{A vector of  DVL-specific infection counts.}
}
\value{
A vector
}
\description{
Gradient of the negative binomial log-likelihood for IUPM from single-dilution assay data (with respect to tau and gamma)
}
