% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion_lrt.R
\name{negbin_loglik_md}
\alias{negbin_loglik_md}
\title{Negative Binomial log-likelihood for IUPM from multiple-dilution assay data}
\usage{
negbin_loglik_md(tau, k, assay_summary)
}
\arguments{
\item{tau}{Vector of DVL-specific IUPM parameters}

\item{k}{Overdispersion parameter (a positive number)}

\item{assay_summary}{Summary of assay data from multiple dilutions}
}
\value{
A scalar value of the negative log-likelihood
}
\description{
Negative Binomial log-likelihood for IUPM from multiple-dilution assay data
}
