% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion_lrt.R
\name{negbin_loglik_sd}
\alias{negbin_loglik_sd}
\title{Negative binomial log-likelihood for IUPM from single-dilution assay data}
\usage{
negbin_loglik_sd(l, k, M, MP, m, Y)
}
\arguments{
\item{l}{Vector of DVL-specific parameters.}

\item{k}{Overdispersion parameter (a positive number)}

\item{M}{Total number of wells originally sequenced with the QVOA.}

\item{MP}{Number of p24-positive wells.}

\item{m}{Number of p24-positive wells that underwent the UDSA.}

\item{y}{A vector of DVL-specific infection counts.}
}
\value{
A scalar
}
\description{
Negative binomial log-likelihood for IUPM from single-dilution assay data
}
