% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_assay_md.R
\name{simulate_assay_md}
\alias{simulate_assay_md}
\title{Simulate multiple-dilution assay data}
\usage{
simulate_assay_md(M, tau, q, u, k = Inf, remove_undetected = T, seed = NULL)
}
\arguments{
\item{M}{Total number of wells at each dilution level (a vector of length D).}

\item{tau}{DVL-specific IUPMs (a vector of length \code{n}). (Note: All elements in \code{tau} must be > 0.)}

\item{q}{Proportions of p24-positive wells that underwent UDSA at each dilution level (a vector of length D).}

\item{u}{Vector of dilution levels (in millions of cells per well).}

\item{k}{Overdispersion parameter (a positive number). Default is Inf, which corresponds to no overdispersion.}

\item{remove_undetected}{Logical, if \code{remove_undetected = TRUE} (the default), then DVL which were not detected in any of the deep sequenced wells across all dilution levels are deleted.}
}
\value{
Named list with the following slots:
\item{assay_summary}{A This summary should contain one row per dilution level and the following columns: M (total number of wells), n (number of distinct viral lineages \[DVL\]), MN (number of p24-negative wells), m (number of deep sequenced wells), Y1,..., Yn (counts of wells positive for DVL i, (i = 1,...,n), and dilutions (dilution levels, in millions of cells per well).}
}
\description{
Simulate multiple-dilution assay data
}
