% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_assay_sd.R
\name{simulate_assay_sd}
\alias{simulate_assay_sd}
\title{Simulate single-dilution assay data}
\usage{
simulate_assay_sd(M, tau, q, u = 1, remove_undetected = TRUE)
}
\arguments{
\item{M}{Total number of wells (a scalar).}

\item{tau}{Mean counts of cells per million infected with each DVL (a vector). (Note: All elements in \code{tau} must be > 0.)}

\item{q}{Proportion of p24-positive wells that underwent UDSA (a scalar between 0 and 1).}

\item{u}{Dilution level in millions of cells per well (a positive scalar). Default is \code{dilution = 1}.}

\item{remove_undetected}{Logical, if \code{remove_undetected = TRUE} (the default), then DVL which were not detected in any of the deep sequenced wells are deleted.}
}
\value{
Named list with the following slots:
\item{any_DVL}{A vector containing overall (any DVL) infection indicators of across the wells.}
\item{DVL_specific}{A matrix of infection indicators with rows and columns representing the DVLs and wells, respectively.}
}
\description{
Simulate single-dilution assay data
}
