% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_SLDeepAssay_sd.R
\name{simulate_SLDeepAssay_sd}
\alias{simulate_SLDeepAssay_sd}
\title{Simulate single-dilution assay data and fit new and existing methods to it.}
\usage{
simulate_SLDeepAssay_sd(
  M,
  tau,
  q,
  u = 1,
  remove_undetected = TRUE,
  seed = NULL
)
}
\arguments{
\item{M}{Total number of wells (a scalar).}

\item{tau}{Mean counts of cells per million infected with each DVL (a vector). (Note: All elements in \code{tau} must be > 0.)}

\item{q}{Fixed proportion of p24-positive wells to be deep sequenced (a scalar between 0 and 1).}

\item{u}{Dilution level in millions of cells per well (a positive scalar). Default is \code{dilution = 1}.}

\item{remove_undetected}{Logical, if \code{remove_undetected = TRUE} (the default), then DVL which were not detected in any of the deep sequenced wells are deleted.}

\item{seed}{(Optional) An integer setting the random seed used to simulate the assay data. Default is \code{seed = NULL}.}
}
\value{
Named list with the following slots:
\item{Assay}{Simulated single-dilution assay data (deep sequenced).}
\item{MLE_woUDSA}{Point estimate, standard error estimate, and 95\% confidence interval for the MLE (without deep sequencing information).}
\item{BCMLE_woUDSA}{Point estimate, standard error estimate, and 95\% confidence interval for the bias-corrected MLE (without deep sequencing information).}
\item{MLE_wUDSA}{Point estimate, standard error estimate, and 95\% confidence interval for the MLE (with deep sequencing information).}
\item{BCMLE_wUDSA}{Point estimate, standard error estimate, and 95\% confidence interval for the bias-corrected MLE (with deep sequencing information).}
\item{Message}{(Optional) Message describing whether the \code{Assay} needed to be re-simulated due to either none of the DVL being detected (\code{Message = 1}) or at least one DVL being detected in all deep-sequenced wells (\code{Message = 2}). Otherwise, \code{Message = 0}.}
\item{Message_Detailed}{(Optional) Count of times the \code{Assay} needed to be re-simulated due to none of the DVL being detected and/or at least one DVL being detected in all deep-sequenced wells. Otherwise, \code{Message_Detailed = NULL}.}
}
\description{
Simulate single-dilution assay data and fit new and existing methods to it.
}
